fixed max_min int input bug

+++ push_swap.h
@@ -51,7 +51,7 @@ void	do_sort_100(t_push **a, t_push **b);
 void	do_sort_500(t_push **a, t_push **b);
 void	sort(t_push **a, t_push **b, int count);
 
-int		ft_atoi(char *str);
+long int		ft_atoi(char *str);
 int		minchr(t_push **a);
 int		maxchr(t_push **a);
 int		numb_amo(t_push **a);
